Использован bulk_create() в PartnerUpdate().

+++ orders_api/api/views.py
@@ -217,6 +217,23 @@ class PartnerUpdate(APIView):
     """
     Класс для обновления прайса от поставщика
     """
+    def _import_product(self, product_data, shop_id):
+        product = Product(name=product_data['name'],
+                          category_id=product_data['category'],
+                          model=product_data['model'],
+                          external_id=product_data['id'],
+                          shop_id=shop_id,
+                          quantity=product_data['quantity'],
+                          price=product_data['price'],
+                          price_rrc=product_data['price_rrc'])
+        return product
+
+    def _import_parameter(self, parameter_list, items_dict, prod):
+        for key, value in items_dict[prod.external_id].items():
+            parameter_list.append(ProductParameter(product_id=prod.id,
+                                                   parameter_id=key,
+                                                   value=value))
+
     def post(self, request, *args, **kwargs):
         if not request.user.is_authenticated:
             return Response({'status': False, 'error': 'Log in required'}, status=status.HTTP_403_FORBIDDEN)
@@ -238,25 +255,31 @@ class PartnerUpdate(APIView):
 
                 shop, _ = Shop.objects.get_or_create(user_id=request.user.id,
                                                      defaults={'name': data['shop'], 'url': url})
+                if shop.name != data['shop']:
+                    return Response({'status': False, 'error': 'В прайсе указано некорректное название магазина!'},
+                                    status=status.HTTP_400_BAD_REQUEST)
                 for category in data['categories']:
                     category_object, _ = Category.objects.get_or_create(id=category['id'], name=category['name'])
                     category_object.shops.add(shop.id)
                     category_object.save()
                 Product.objects.filter(shop_id=shop.id).delete()
+                product_items = []
+                parameter_items = {}
+                parameter_list = []
                 for item in data['goods']:
-                    product = Product.objects.create(name=item['name'],
-                                                     category_id=item['category'],
-                                                     model=item['model'],
-                                                     external_id=item['id'],
-                                                     shop_id=shop.id,
-                                                     quantity=item['quantity'],
-                                                     price=item['price'],
-                                                     price_rrc=item['price_rrc'])
+                    product_items.append(self._import_product(item, shop.id))
+                    parameter_items[item['id']] = {}
                     for name, value in item['parameters'].items():
                         parameter, _ = Parameter.objects.get_or_create(name=name)
-                        ProductParameter.objects.create(product_id=product.id,
-                                                        parameter_id=parameter.id,
-                                                        value=value)
+                        parameter_items[item['id']].update({parameter.id: value})
+
+                list_of_products = Product.objects.bulk_create(product_items)
+
+                for product in list_of_products:
+                    self._import_parameter(parameter_list, parameter_items, product)
+
+                ProductParameter.objects.bulk_create(parameter_list)
+
                 return Response({'status': True})
 
         return Response({'status': False, 'error': 'Не указаны все необходимые аргументы'},
